% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvars.R
\docType{data}
\name{col_name}
\alias{col_name}
\title{The Korean and English names of the column names}
\format{
An object of class \code{data.frame} with 174 rows and 3 columns.
}
\usage{
col_name
}
\description{
The Korean and English names of the column names
}
\keyword{datasets}
