% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globalvars.R
\docType{data}
\name{nfi_donghae}
\alias{nfi_donghae}
\title{National Forest Inventory data for Donghae-si, Gangwon-do, Republic of Korea for testing the function}
\format{
An object of class \code{list} of length 3.
}
\usage{
nfi_donghae
}
\description{
National Forest Inventory data for Donghae-si, Gangwon-do, Republic of Korea for testing the function
}
\keyword{datasets}
